feat(`uninstall.sh`): ask user for confirmation

+++ scripts/uninstall.sh
@@ -10,19 +10,27 @@ if [ "$EUID" -ne 0 ]
   exit
 fi
 
-echo "Stopping reNgine"
-docker stop rengine-web-1 rengine-db-1 rengine-celery-1 rengine-celery-beat-1 rengine-redis-1 rengine-tor-1 rengine-proxy-1
+read -p "This action will remove all containers, volumes and networks of reNgine. Do you want to continue? [y/n] " -n 1 -r
+echo ""
 
-echo "Removing all containers related to reNgine"
-docker rm rengine-web-1 rengine-db-1 rengine-celery-1 rengine-celery-beat-1 rengine-redis-1 rengine-tor-1 rengine-proxy-1
-echo "Removed all containers"
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  echo "Stopping reNgine"
+  docker stop rengine-web-1 rengine-db-1 rengine-celery-1 rengine-celery-beat-1 rengine-redis-1 rengine-tor-1 rengine-proxy-1
 
-echo "Removing all volumes related to reNgine"
-docker volume rm rengine_gf_patterns rengine_github_repos rengine_nuclei_templates rengine_postgres_data rengine_scan_results rengine_tool_config
-echo "Removed all volumes"
+  echo "Removing all containers related to reNgine"
+  docker rm rengine-web-1 rengine-db-1 rengine-celery-1 rengine-celery-beat-1 rengine-redis-1 rengine-tor-1 rengine-proxy-1
+  echo "Removed all containers"
 
-echo "Removing all networks related to reNgine"
-docker network rm rengine_rengine_network rengine_default
+  echo "Removing all volumes related to reNgine"
+  docker volume rm rengine_gf_patterns rengine_github_repos rengine_nuclei_templates rengine_postgres_data rengine_scan_results rengine_tool_config
+  echo "Removed all volumes"
+
+  echo "Removing all networks related to reNgine"
+  docker network rm rengine_rengine_network rengine_default
+else
+  exit 1
+fi
 
 read -p "Do you want to remove Docker images related to reNgine? [y/n] " -n 1 -r
 echo ""
